fix: events of view can be register by need
