+++ packages/vgrammar-core/src/parse/event.ts
@@ -11,21 +11,21 @@ import { isMarkType } from '../graph/util/graphic';
 import { EVENT_SOURCE_VIEW, EVENT_SOURCE_WINDOW } from '../view/constants';
 
 export const generateFilterByMark = (evtSpec: Partial<ParsedViewEventSpec | ParsedWindowEventSpec>) => {
-  if (isNil(evtSpec.markId)) {
+  if (!isNil(evtSpec.markId)) {
     return (el: IElement) => {
       return el && el.mark.id() === evtSpec.markId;
     };
   }
 
-  if (isNil(evtSpec.markName)) {
+  if (!isNil(evtSpec.markName)) {
     return (el: IElement) => {
       return el && el.mark.name() === evtSpec.markName;
     };
   }
 
-  if (isNil(evtSpec.type)) {
+  if (!isNil(evtSpec.markType)) {
     return (el: IElement) => {
-      return el && el.mark.markType === evtSpec.type;
+      return el && el.mark.markType === evtSpec.markType;
     };
   }
