+++ packages/vgrammar-core/src/view/View.ts
@@ -153,7 +153,6 @@ export default class View extends EventEmitter implements IView {
   private _layoutMarks?: IMark[];
 
   private _background?: IColor;
-  private _eventCache: Record<string, () => void>;
   /** 当前是否存在增量渲染元素 */
   private _progressiveMarks?: IMark[];
   private _progressiveRafId?: number;
@@ -534,7 +533,7 @@ export default class View extends EventEmitter implements IView {
 
     if (spec.events && spec.events.length) {
       spec.events.forEach(eventConfig => {
-        this.event(eventConfig);
+        (this as any).event?.(eventConfig);
       });
     }
 
@@ -1121,170 +1120,6 @@ export default class View extends EventEmitter implements IView {
     }
 
     return false;
-  }
-
-  // --- Event ---
-
-  private bindEvents(eventSpec: BaseEventSpec) {
-    if (this._eventConfig.disable) {
-      return;
-    }
-    const { type: evtType, filter, callback, throttle, debounce, consume, target, dependency } = eventSpec;
-    const eventSelector = parseEventSelector(evtType);
-
-    if (!eventSelector) {
-      return;
-    }
-    const { source, type } = eventSelector;
-
-    const markFilter = generateFilterByMark(eventSelector);
-    const targetSignals =
-      Array.isArray(target) && target.length
-        ? target.map(entry => {
-            return {
-              signal: this.getSignalById(entry.target),
-              callback: entry.callback
-            };
-          })
-        : [
-            {
-              signal: isString(target) ? this.getSignalById(target) : null,
-              callback
-            }
-          ];
-    const validateSignals = targetSignals.filter(entry => entry.signal || entry.callback);
-    const refs = parseReference(dependency, this);
-
-    const send = parseHandler(
-      (evt?: any, element?: IElement) => {
-        const needPreventDefault =
-          (source === EVENT_SOURCE_VIEW && prevent(this._eventConfig, type)) ||
-          (consume && (evt.cancelable === undefined || evt.cancelable));
-
-        if (source === EVENT_SOURCE_WINDOW) {
-          evt = getExtendedEvents(this, evt, element, type, EVENT_SOURCE_WINDOW);
-        }
-
-        let hasCommitted = false;
-
-        if ((!filter || filter(evt)) && (!markFilter || markFilter(element)) && validateSignals.length) {
-          const params = refs.reduce((params, ref) => {
-            params[ref.id()] = ref.output();
-            return params;
-          }, {});
-          validateSignals.forEach(entry => {
-            if (entry.callback && entry.signal) {
-              const changed = entry.signal.set(entry.callback(evt, params));
-
-              if (changed) {
-                this.commit(entry.signal);
-                hasCommitted = true;
-              }
-            } else if (entry.callback) {
-              entry.callback(evt, params);
-            } else {
-              this.commit(entry.signal);
-              hasCommitted = true;
-            }
-          });
-        }
-
-        if (needPreventDefault) {
-          evt.preventDefault();
-        }
-
-        if (consume) {
-          evt.stopPropagation();
-        }
-
-        if (hasCommitted) {
-          this.run();
-        }
-      },
-      { throttle, debounce }
-    );
-
-    if (source === EVENT_SOURCE_VIEW) {
-      if (permit(this._eventConfig, EVENT_SOURCE_VIEW, type)) {
-        // send traps errors, so use {trap: false} option
-        this.addEventListener(type, send, NO_TRAP);
-
-        return () => {
-          this.removeEventListener(type, send);
-        };
-      }
-    } else if (source === EVENT_SOURCE_WINDOW) {
-      vglobal.addEventListener(type, send);
-      this._eventListeners.push({
-        type: type,
-        source: vglobal,
-        handler: send
-      });
-
-      return () => {
-        vglobal.removeEventListener(type, send);
-
-        const index = this._eventListeners.findIndex((entry: any) => {
-          return entry.type === type && entry.source === vglobal && entry.handler === send;
-        });
-
-        if (index >= 0) {
-          this._eventListeners.splice(index, 1);
-        }
-      };
-    }
-  }
-
-  event(eventSpec: EventSpec) {
-    if ('between' in eventSpec) {
-      const [starEvent, endEvent] = eventSpec.between;
-      // FIXME between需要生成唯一ID
-      const id = `${starEvent.type}-${eventSpec.type}-${endEvent.type}`;
-
-      let unbindEndEvent: any;
-
-      this.bindEvents(
-        Object.assign({}, starEvent, {
-          callback: () => {
-            if (!this._eventCache) {
-              this._eventCache = {};
-            }
-            // 中间的事件绑定
-            if (!this._eventCache[id]) {
-              const unbindEvent = this.bindEvents(eventSpec);
-              this._eventCache[id] = unbindEvent;
-            }
-            if (!unbindEndEvent) {
-              // 结束的事件绑定
-              unbindEndEvent = this.bindEvents(
-                Object.assign({}, endEvent, {
-                  callback: () => {
-                    if (this._eventCache[id]) {
-                      this._eventCache[id]();
-                      this._eventCache[id] = null;
-                    }
-                  }
-                })
-              );
-            }
-          }
-        })
-      );
-    } else if ('merge' in eventSpec) {
-      eventSpec.merge.forEach(entry => {
-        const singleEvent: Partial<BaseEventSpec> = Object.assign({}, eventSpec);
-
-        if (isString(entry)) {
-          singleEvent.type = entry;
-        } else if (isObject(entry)) {
-          Object.assign(singleEvent, entry);
-        }
-        singleEvent.debounce = 50;
-        this.bindEvents(singleEvent as BaseEventSpec);
-      });
-    } else {
-      this.bindEvents(eventSpec);
-    }
   }
 
   interaction(type: string, spec: Partial<InteractionSpec>) {
